Add AppProvider (#48)
- Adds `AppProvider` – this is an abstraction that encapsulates the
application creation and respective setup
- There are two implementations:
* `DefaultAppProvider` – is the default application instance and should
be the one used for normal application instantiation
* `TestAppProvider` – provides application instance given a
`TestingModule` – providing any other module would result in a rejected
`Promise`
- These two implementation share the same setup steps – this allows us
to write tests that will use the same configuration applied in the
production instance. The configuration steps are set under
`DEFAULT_CONFIGURATION`. This configuration is provided to the instances
so that it can be changed (e.g. a testing instance might want to remove
a configuration step)