Add RouteLoggerInterceptor (#63)
- Adds `RouteLoggerInterceptor` – a NestJS interceptor which logs the
requests that target each route
- Removes the request logging from the `LoggerMiddleware`
- The main reason for this change is that Middlewares are executed
before the route handlers – this means that we lack context on which
handler is going to be called at this stage. Interceptors however, have
an `ExecutionContext` tied to it. Which means that at this point we have
more information on the route that is going to be triggered.
- Removes the logging of `contentLength` and `contentType` since these
were empty in the middleware (this needs to be reviewed internally, i.e.
if and how we want to log these values)

For example if we trigger the relay route `GET /v1/relay/1/0x0`:

```
console.log(req.route.path)
"/*" // LoggerMiddleware
"/v1/relay/:chainId/:address" // RouteLoggerInterceptor
```